+++ packages/sdk/money-like-to-node/src/toNodeEntities/updateAssetInfo.ts
@@ -1,4 +1,4 @@
-import { type UpdateAssetInfoTransaction } from '@decentralchain/ts-types';
+import { type UpdateAssetInfoTransaction } from '@decentralchain/types';
 import { type TYPES } from '../constants/index.js';
 import { factory } from '../core/factory.js';
 import { type TWithPartialFee } from '../types/index.js';